feat: create modal and reacfored code
